% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available-package.R
\docType{package}
\name{available-package}
\alias{available-package}
\title{available: Check if the Title of a Package is Available, Appropriate and Interesting}
\description{
Check if a given package name is available to use. It checks the name's validity. Checks if it is used on 'GitHub', 'CRAN' and 'Bioconductor'. Checks for unintended meanings by querying 'Wiktionary' and Wikipedia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-lib/available}
  \item \url{https://r-lib.github.io/available/}
  \item Report bugs at \url{https://github.com/r-lib/available/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Carl Ganz
  \item Jim Hester
  \item Molly Lewis
  \item Rachael Tatman
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
}

}
\keyword{internal}
