% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{available_github}
\alias{available_github}
\title{See if a name is available on github}
\usage{
available_github(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
See if a name is available on github
}
