% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran.R
\name{available_on_cran}
\alias{available_on_cran}
\title{See if a name is available on CRAN}
\usage{
available_on_cran(name, ...)
}
\arguments{
\item{name}{Name of package to search}

\item{...}{Additional arguments passed to [utils::available.packages()].}
}
\description{
See if a name is available on CRAN
}
