% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_words.R
\name{check_bad_words}
\alias{check_bad_words}
\title{Check for bad words in name}
\usage{
check_bad_words(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
Check for bad words in name
}
