% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namr.R
\name{common_suffixes}
\alias{common_suffixes}
\title{Add common suffixes}
\usage{
common_suffixes(title, name)
}
\arguments{
\item{title}{the package title or discription}

\item{name}{the single word that will be appended to}
}
\value{
a single word with affix, if applicable
}
\description{
Search a title for common terms (plot, vis..., viz..., markdown) and apply
appropriate affixes to a given word as appliable.
}
\keyword{internal}
