% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{create}
\alias{create}
\title{Check a new package name and possibly create it}
\usage{
create(name, ...)
}
\arguments{
\item{name}{Name of package to search}

\item{...}{Additional arguments passed to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\description{
Check a new package name and possibly create it
}
