% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namr.R
\name{find_acronym}
\alias{find_acronym}
\title{Function that finds and returns the first acronym (all caps) in a text string}
\usage{
find_acronym(title)
}
\arguments{
\item{title}{package title or description}
}
\value{
a single acronym, if present
}
\description{
Function that finds and returns the first acronym (all caps) in a text string
}
\keyword{internal}
