% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad_words.R
\name{get_bad_words}
\alias{get_bad_words}
\title{Check for bad words in name}
\usage{
get_bad_words(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
Check for bad words in name
}
\seealso{
See \url{https://github.com/web-mech/badwords}
}
