% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_urls.R
\name{get_sentiment}
\alias{get_sentiment}
\title{Open wikipedia page and abbreviations.com page}
\usage{
get_sentiment(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
Open wikipedia page and abbreviations.com page
}
