% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tribe_names.R
\name{get_tribe_names}
\alias{get_tribe_names}
\title{Check for tribe names in pkg name}
\usage{
get_tribe_names(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
Check for tribe names in pkg name
}
\seealso{
See \url{https://www.bia.gov/tribal-leaders-directory}
}
