% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban_data.R
\name{get_urban_data}
\alias{get_urban_data}
\title{get urban dictionary definitions and tags}
\usage{
get_urban_data(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
get urban dictionary definitions and tags
}
\note{
Urban Dictionary can contain potentially offensive results
}
