% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki.R
\name{get_wikipidia}
\alias{get_wikipidia}
\alias{get_wiktionary}
\title{Open wikipedia page and abbreviations.com page}
\usage{
get_wikipidia(name)

get_wiktionary(name)
}
\arguments{
\item{name}{Name of package to search}
}
\description{
Open wikipedia page and abbreviations.com page
}
