% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namr.R
\name{make_spelling_rlike}
\alias{make_spelling_rlike}
\title{Spelling transformations}
\usage{
make_spelling_rlike(word)
}
\arguments{
\item{word}{a single word to make more rlike}
}
\value{
a single word with a spelling transformation
}
\description{
This function takes in a single word and applies spelling transformations to
make it more "r-like" and easier to Google
}
\keyword{internal}
