% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namr.R
\name{namr}
\alias{namr}
\title{Suggest package name}
\usage{
namr(title, acronym = F, verb = F, ...)
}
\arguments{
\item{title}{the package title or discription}

\item{acronym}{whether to include an acronym (if there is one) in the title}

\item{verb}{whether to prioritize using a verb in the package title}
}
\value{
a single word to use as a package title
}
\description{
Suggest package name
}
\keyword{internal}
