% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namr.R
\name{pick_word_from_title}
\alias{pick_word_from_title}
\title{Pick word from title}
\usage{
pick_word_from_title(title, verb = F)
}
\arguments{
\item{title}{text string to pick word from. Package title or description.}

\item{verb}{whether you would like to prioritize returning a verb}
}
\value{
a single word from the title
}
\description{
picks a single (hopefully informative) word from the provided title or package description
}
\keyword{internal}
