% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist.R
\name{pkg_name_dist}
\alias{pkg_name_dist}
\title{Find five closest package names in terms of string distance}
\usage{
pkg_name_dist(name, pkgs)
}
\arguments{
\item{name}{name of package}

\item{pkgs}{packages to compare with}
}
\description{
Find five closest package names in terms of string distance
}
