% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{suggest}
\alias{suggest}
\title{Suggest a title based for a development package}
\usage{
suggest(path = ".", title = NULL)
}
\arguments{
\item{path}{Path to a existing package to extract the title from.}

\item{title}{title string to search.}
}
\description{
If the package you are using already has a title, simply pass the path to
the package root in `path`. Otherwise use `title` to specify a potential
title.
}
