% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\title{See if a name is available}
\usage{
available(name, ...)
}
\arguments{
\item{name}{Name of package to search}

\item{...}{Additional arguments passed to [utils::available.packages()].}
}
\description{
Searches performed
- Valid package name
- Already taken on CRAN
- Positive or negative sentiment
- Urban Dictionary
}
