% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioc.R, R/cran.R
\name{available_on_bioc}
\alias{available_on_bioc}
\alias{available_on_cran}
\title{See if a name is available on CRAN}
\usage{
available_on_bioc(name, repos = NULL, ...)

available_on_cran(name, repos = default_cran_repos, ...)
}
\arguments{
\item{name}{Name of package to search}

\item{repos}{
    character vector, the base URL(s) of the repositories to use.
  }

\item{...}{Additional arguments passed to \code{\link[utils:available.packages]{utils::available.packages()}}.}
}
\description{
See if a name is available on CRAN
}
